deployment/ecr/gaab-strands-agent/pyproject.toml
[project]
name = "gaab-strands-agent"
version = "4.0.0"
description = "GAAB Strands Agent Runtime for Amazon Bedrock AgentCore"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "Apache-2.0" }
authors = [{ name = "Amazon Web Services" }]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "setuptools>=70.0.0",
    "pip>=25.0",
    "wheel>=0.42.0",

    # AWS SDK
    "boto3>=1.35.0",

    # Strands SDK with OpenTelemetry support
    "strands-agents[otel]>=1.10.0",
    "strands-agents-tools>=0.2.9",

    # bedrock-agentcore
    "bedrock-agentcore>=0.1.5",

    # OpenTelemetry for observability
    "aws-opentelemetry-distro>=0.12.1",

    # Pydantic for data validation and parsing
    "pydantic>=2.0.0",

    # Shared library
    "gaab-strands-common",
]

[tool.uv.sources]
gaab-strands-common = { path = "../gaab-strands-common", editable = true }

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "isort>=5.12.0",
    "mypy>=1.8.0",
]

test = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "moto>=5.0.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "isort>=5.12.0",
    "mypy>=1.8.0",
    "moto>=5.0.0",
    "pytest-mock>=3.12.0",
]

[tool.black]
line-length = 100
target-version = ['py313']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["test"]
pythonpath = ["."]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--import-mode=importlib"
